actualizacionn

+++ src/main/java/com/senati/senati_web_mallqui/model/User.java
@@ -14,7 +14,7 @@ public class User {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
